Simplify `CorsTestFixture<TStartup>` and ensure `EchoMiddleware` is _not_ firing and forgetting - PR comments
